requirements.txt
Django==1.7.1
Fabric==1.10.1
dj-database-url==0.3.0
django-debug-toolbar==1.2.2
django-dotenv==1.3.0
django-getenv==1.3.1
djangorestframework==3.0.3
ecdsa==0.11
gunicorn==19.1.1
paramiko==1.15.2
pycrypto==2.6.1
sqlparse==0.1.14
whitenoise==1.0.6
wsgiref==0.1.2
